#41 Refactor: 초기 데이터 위치 변경
